Move WarningMessage to flat structure
