+++ packages/components/src/components/forms/index.ts
@@ -21,7 +21,6 @@ export * from './VerifyingIndicator'
 export * from './ValidIndicator'
 export * from './CheckboxGroup/CheckboxGroup'
 export * from './ImageUploader'
-export * from './WarningMessage'
 export * from './PDFViewer/PDFViewer'
 export * from './ErrorMessage'
 export * from './InputField/InputLabel'
