+++ packages/login/src/views/resetCredentialsForm/updatePasswordForm.tsx
@@ -16,11 +16,8 @@ import {
 } from '@login/login/actions'
 import { authApi } from '@login/utils/authApi'
 import { PrimaryButton } from '@opencrvs/components/lib/buttons'
-import {
-  InputField,
-  TextInput,
-  WarningMessage
-} from '@opencrvs/components/lib/forms'
+import { InputField, TextInput } from '@opencrvs/components/lib/forms'
+import { WarningMessage } from '@opencrvs/components/lib/WarningMessage'
 import { TickOff, TickOn } from '@opencrvs/components/lib/icons'
 import { ActionPageLight } from '@opencrvs/components/lib/interface'
 import * as React from 'react'
